requirements.txt
numpy==1.26.4
scikit-learn==1.6.1
opencv-python==4.11.0.86
facenet-pytorch==2.6.0
torch==2.2.2
mediapipe==0.10.20
insightface==0.7.3
faiss-cpu==1.10.0
matplotlib==3.10.0